Fixes the failing tests for the visualizations module.
This was done by factoring out the code that actually generates and
saves the output from each function in pyret.visualizations into the
utils module. There is now an executable `create_baseline_images.py`,
which is used to generate the baseline images via the *same code* that
is used to generate the test images. I have verified visually the images
generated, but the fact that this commit passes the tests is more or
less meaningless because it's the exact same code used to generate the
test and baseline images. But re-generating these baseline images should
only need to be done in rare circumstances, such as the change in
matplotlib that caused the break in pyret addressed by this commit. So
future commits that change pyret.visualizations will actually be tested
by these tests.

Long story short, the tests are still useful, but don't run
`create_baseline_images.py` unless you really need to.