Changes testing strategy for the visualization module.
Previous versions tested the module by comparing images. The tests
generated an image, which was compared pixel-to-pixel with a baseline
version saved in the repository. This is complicated, error-prone, and
difficult to maintain.

This commit now tests the module by calling the functions with simple
and known data, and then verifying that the data in the actual plot
objects rendered matches our expectations based on the inputs.