Adds option to all visualizations which display spatial filters to (#95)
* Adds option to all visualizations which display spatial filters to
accept a dx keyword argument. This gives the spatial scale of a single
pixel in the passed linear filter, which is used to set the axis extent
accordingly. This allows users to see spatial filters in actual spatial
units. The parameter defaults to 1.0 (plot in pixel coordinates), so the
behavior for older code is the same.

* Adds changes suggested by Niru. Some line-wrapping for PEP8 compliance.
Adds try/except clauses around routines in visualizations.plot_cells
which may raise, so that one STA not being decomposable doesn't prevent
the rest from being plotted.

* bumps version to 0.5.7