Fixes #96
This commit fixes a bug in `linear_response()`. That method was written
to accept a linear filter, but it previously only took an STA or
reverse-correlation. It ignored the initial (but still valid) portion of
the response of a filter to a stimulus. The new method, with improved
docstring, explains more clearly what is expected, and indicates that
the new output should have the same shape as the stimulus passed in. The
quickstart was also updated.

The docstring for `stimulustools.slicestim` was also incorrect, stating
that it returned an array of a different shape than the one actually
returned.

Also removes the old `rolling_window` function, and bumps the version
number.