Removed stray print from testing.
Also, noting in this commit message that a previous one (#d503279) fixed
a bug in `filtertool.sta`. That method only counted a set of frames in
the output once, regardless of the number of spikes in the time bin.
Multiple spikes should count for multiple copies of that same stimulus
chunk.