Bug fix in linear response method. (#97)
* Fixes #96

This commit fixes a bug in `linear_response()`. That method was written
to accept a linear filter, but it previously only took an STA or
reverse-correlation. It ignored the initial (but still valid) portion of
the response of a filter to a stimulus. The new method, with improved
docstring, explains more clearly what is expected, and indicates that
the new output should have the same shape as the stimulus passed in. The
quickstart was also updated.

The docstring for `stimulustools.slicestim` was also incorrect, stating
that it returned an array of a different shape than the one actually
returned.

Also removes the old `rolling_window` function, and bumps the version
number.

* updates changelog for v0.6.0

* Removed stray print from testing.

Also, noting in this commit message that a previous one (#d503279) fixed
a bug in `filtertool.sta`. That method only counted a set of frames in
the output once, regardless of the number of spikes in the time bin.
Multiple spikes should count for multiple copies of that same stimulus
chunk.

* fixes python3.5+-specific tuple unpacking error

* Correctly handles acausal points in filter in new version of
``linear_response``, including error handling.